build(gamify): small code refactoring + tweaks (#237)
* configure CODEOWNERS

* chore: refactor 'gamify-api-*' into 'gamify-*

* chore: configure eslint boundaries

* chore: tweaks
