+++ apps/gamify/api/project.json
@@ -48,5 +48,5 @@
       }
     }
   },
-  "tags": []
+  "tags": ["type:app", "scope:gamify"]
 }
