--- libs/gamify-api/core/project.json
@@ -1,22 +1,0 @@
-{
-  "root": "libs/gamify-api/core",
-  "sourceRoot": "libs/gamify-api/core/src",
-  "targets": {
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["libs/gamify-api/core/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/gamify-api/core"],
-      "options": {
-        "jestConfig": "libs/gamify-api/core/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
