--- libs/gamify-api/core/src/index.ts
@@ -1,3 +1,0 @@
-export * from './lib/gamify-api-core.module';
-export * from './lib/enums';
-export * from './lib/interfaces';
