+++ libs/gamify/apps/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'gamify-api-core',
+  displayName: 'gamify-apps',
   preset: '../../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -11,5 +11,5 @@ module.exports = {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../../coverage/libs/gamify-api/core',
+  coverageDirectory: '../../../coverage/libs/gamify/apps',
 };
