+++ libs/gamify/apps/project.json
@@ -1,0 +1,22 @@
+{
+  "root": "libs/gamify/apps",
+  "sourceRoot": "libs/gamify/apps/src",
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/gamify/apps/**/*.ts"]
+      }
+    },
+    "test": {
+      "executor": "@nrwl/jest:jest",
+      "outputs": ["coverage/libs/gamify/apps"],
+      "options": {
+        "jestConfig": "libs/gamify/apps/jest.config.js",
+        "passWithNoTests": true
+      }
+    }
+  },
+  "tags": ["type:feature", "scope:gamify"]
+}
