+++ libs/gamify/core/.eslintrc.json
