+++ libs/gamify/core/project.json
@@ -1,22 +1,22 @@
 {
-  "root": "libs/gamify-api/apps",
-  "sourceRoot": "libs/gamify-api/apps/src",
+  "root": "libs/gamify/core",
+  "sourceRoot": "libs/gamify/core/src",
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/gamify-api/apps/**/*.ts"]
+        "lintFilePatterns": ["libs/gamify/core/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/libs/gamify-api/apps"],
+      "outputs": ["coverage/libs/gamify/core"],
       "options": {
-        "jestConfig": "libs/gamify-api/apps/jest.config.js",
+        "jestConfig": "libs/gamify/core/jest.config.js",
         "passWithNoTests": true
       }
     }
   },
-  "tags": []
+  "tags": ["type:util", "scope:gamify"]
 }
