+++ libs/gamify/core/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/gamify-core.module';
+export * from './lib/enums';
+export * from './lib/interfaces';
