+++ libs/gamify/core/src/lib/enums/application-type.enum.ts
