+++ libs/gamify/core/src/lib/interfaces/app.interface.ts
