+++ libs/gamify/core/src/lib/interfaces/index.ts
